more tfjs tidying up and electron integration
