+++ js/webcomponents/bisweb_mainviewerapplication.js
@@ -1206,16 +1206,23 @@ class ViewerApplicationElement extends HTMLElement {
 
         let show=force;
 
-        Promise.all( [ 
+        let p=[ 
             userPreferences.safeGetImageOrientationOnLoad(),
             userPreferences.safeGetItem('showwelcome'),
             webutil.aboutText(),
-            idb.get('mode')
-        ]).then( (lst) => {
+        ];
+
+        if (!webutil.inElectronApp()) {
+            p.push( idb.get('mode'));
+        }
+
+        Promise.all(p).then( (lst) => {
             let forceorient=lst[0];
             let firsttime=lst[1];
             let msg=lst[2];
 
+            lst[3]=lst[3] || '';
+            
             let offline=false;
             if (lst[3].indexOf('offline')>=0)
                 offline=true;
