+++ package.json
@@ -1,7 +1,7 @@
 {
     "private": true,
-    "name": "bioimagesuiteweb_wasm",
-    "version": "0.9.1",
+    "name": "bioimagesuiteweb",
+    "version": "1.1.0",
     "description": "An Implementation of BioImage Suite in Javascript and WebAssembly",
     "homepage": "https://github.com/bioimagesuiteweb/bisweb",
     "scripts": {
